notifications: keep type label in UI layer, single KNOWN_TYPES
Address review feedback on #16588:
- Move typeLabelFor back into NotificationsViewModel (the Context it
  needs already lives there), and remove it plus the @ApplicationContext
  Context from the repository interface/impl so the data layer no longer
  holds an Android dependency solely for string resolution.
- Collapse the duplicated KNOWN_TYPES into a single definition on the
  NotificationsRepository interface companion; the ViewModel now reads
  the repository's set, so the classifier guard and the grouping bucket
  can no longer drift.
- Drop the per-PR version bump; automerge owns the version during the
  drain and carrying it conflicts with every other bumped PR.
- Keep resolveType in the repository; update tests accordingly.
