+++ app/build.gradle
@@ -10,8 +10,8 @@ android {
         applicationId "org.ole.planet.myplanet"
         minSdk = 26
         targetSdk = 36
-        versionCode = 6768
-        versionName = "0.67.68"
+        versionCode = 6767
+        versionName = "0.67.67"
         ndkVersion = '26.3.11579264'
         vectorDrawables.useSupportLibrary = true
     }
