+++ app/src/main/java/org/ole/planet/myplanet/repository/NotificationsRepositoryImpl.kt
@@ -1,16 +1,13 @@
 package org.ole.planet.myplanet.repository
 
-import android.content.Context
 import com.google.gson.JsonArray
 import com.google.gson.JsonObject
 import dagger.Lazy
-import dagger.hilt.android.qualifiers.ApplicationContext
 import java.util.Calendar
 import java.util.Date
 import java.util.Locale
 import java.util.UUID
 import javax.inject.Inject
-import org.ole.planet.myplanet.R
 import org.ole.planet.myplanet.data.room.dao.ExamDao
 import org.ole.planet.myplanet.data.room.dao.NotificationDao
 import org.ole.planet.myplanet.data.room.dao.TeamNotificationDao
@@ -26,7 +23,6 @@ import org.ole.planet.myplanet.utils.TimeProvider
 private const val STORAGE_WARNING_AVAILABLE_PERCENT = 10
 
 class NotificationsRepositoryImpl @Inject constructor(
-    @ApplicationContext private val context: Context,
     private val userRepository: Lazy<UserRepository>,
     private val teamsRepository: Lazy<TeamsNotificationsRepository>,
     private val timeProvider: TimeProvider,
@@ -348,7 +344,7 @@ class NotificationsRepositoryImpl @Inject constructor(
     }
 
     override fun resolveType(type: String, message: String, subType: String?): String {
-        if (type.lowercase(Locale.ROOT) in KNOWN_TYPES) return type.lowercase(Locale.ROOT)
+        if (type.lowercase(Locale.ROOT) in NotificationsRepository.KNOWN_TYPES) return type.lowercase(Locale.ROOT)
         val lower = message.lowercase(Locale.ROOT)
         // Raw server type "team" covers every team-related event (message/request/added/rejected/removed) in
         // whatever language the server rendered the message in, so classify structurally first and only fall
@@ -375,17 +371,6 @@ class NotificationsRepositoryImpl @Inject constructor(
             lower.contains("resource") -> "resource"
             else -> "notification"
         }
-    }
-
-    override fun typeLabelFor(type: String): String = when (type.lowercase(Locale.ROOT)) {
-        "join_request" -> context.getString(R.string.notif_group_join_requests)
-        "team_join" -> context.getString(R.string.notif_group_team_updates)
-        "task" -> context.getString(R.string.tasks)
-        "chat" -> context.getString(R.string.notif_group_new_voices)
-        "voice_reply" -> context.getString(R.string.notif_group_voice_replies)
-        "resource" -> context.getString(R.string.resources)
-        "storage" -> context.getString(R.string.notification_group_system)
-        else -> context.getString(R.string.notification_group_other)
     }
 
     private fun parseNotification(doc: JsonObject): AppNotification? {
@@ -480,9 +465,5 @@ class NotificationsRepositoryImpl @Inject constructor(
             }
         }
         notificationDao.upsertAll(parsedList)
-    }
-
-    companion object {
-        val KNOWN_TYPES = setOf("join_request", "team_join", "task", "chat", "voice_reply", "resource", "storage")
     }
 }
